==> notebooks/1 gather_data.ipynb <==
# Data Gathering

All data in this project is pulled from the [PokéAPI](https://pokeapi.co/). Full credit and huge thanks to them.

# dependencies 
import requests
import json

type_url = "https://pokeapi.co/api/v2/type/"
species_url = "https://pokeapi.co/api/v2/pokemon-species/"
pkmn_url = "https://pokeapi.co/api/v2/pokemon/"

## Gathering Pokemon Types

I want to get all of the types. This API includes two types, Shadow and Unknown, that I'm not interested in, so I'm going to limit the results to the 18 base types I know are in Pokémon games.

# get list of types, extract the api urls
types = requests.get(type_url, params={"limit": 18}).json()

types_urls = [types['results'][type]['url'] for type in range(len(types['results']))]

# extract each type and add it to a list
types = []

for url in types_urls:
    try:
        types.append(requests.get(url).json())
    except:
        print('Skipping. Trying next type')


# export types to a file
with open('API_Data/types.json', 'w') as out_file:
    json.dump(types, out_file, indent=2)



## Gathering Pokémon Species

This dataset has "Pokémon Species" and "Pokémon" datasets. The species dataset contains one entry for each species of Pokémon with keys that define any variants. The Pokémon dataset has a separate entry for each Pokémon, without grouping by series. The species set is smaller and gives you a list of forms. The Pokémon dataset lists the types of each form. I want both datasets.

species = requests.get(species_url).json()

species_len = species['count']

species = []

for id in range(1,species_len):
    try:
        species.append(requests.get(f'{species_url}{id}').json())
    except:
        print(f'Skipping pkmn {id}. Trying next pkmn.')

# export pokemon species to a file
with open('API_Data/pkmn_species.json', 'w') as out_file:
    json.dump(species, out_file, indent=2)

## Gathering Pokémon

In this dataset the base forms of Pokémon are numbered 1-905, and then alternate forms (such as region-specific forms) are listed in 10001-10249. Alternate forms often have different types so I'm going to get them all in two passes.

pkmn = []

for id in range(1,906):
    try:
        pkmn.append(requests.get(f'{pkmn_url}{id}').json())
    except:
        print(f'Skipping pkmn {id}. Trying next pkmn.')


for id in range(10001,10250):
    try:
        pkmn.append(requests.get(f'{pkmn_url}{id}').json())
    except:
        print(f'Skipping pkmn {id}. Trying next pkmn.')

# export pokemon to a file
with open('API_Data/pkmn.json', 'w') as out_file:
    json.dump(pkmn, out_file, indent=2)
